Move from logic.autosave to logic.autosaveandbackup
