+++ src/main/java/net/sf/jabref/gui/JabRefFrame.java
@@ -108,8 +108,8 @@ import net.sf.jabref.gui.search.GlobalSearchBar;
 import net.sf.jabref.gui.util.WindowLocation;
 import net.sf.jabref.gui.worker.MarkEntriesAction;
 import net.sf.jabref.logic.CustomEntryTypesManager;
-import net.sf.jabref.logic.autosave.AutosaveManager;
-import net.sf.jabref.logic.autosave.BackupManager;
+import net.sf.jabref.logic.autosaveandbackup.AutosaveManager;
+import net.sf.jabref.logic.autosaveandbackup.BackupManager;
 import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.logic.importer.OutputPrinter;
 import net.sf.jabref.logic.importer.ParserResult;
