+++ src/main/java/net/sf/jabref/gui/autosave/BackupUIManager.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 import javax.swing.JFrame;
 import javax.swing.JOptionPane;
 
-import net.sf.jabref.logic.autosave.BackupManager;
+import net.sf.jabref.logic.autosaveandbackup.BackupManager;
 import net.sf.jabref.logic.l10n.Localization;
 
 /**
