+++ src/main/java/net/sf/jabref/gui/exporter/SaveDatabaseAction.java
@@ -24,8 +24,8 @@ import net.sf.jabref.gui.autosave.AutosaveUIManager;
 import net.sf.jabref.gui.worker.AbstractWorker;
 import net.sf.jabref.gui.worker.CallBack;
 import net.sf.jabref.gui.worker.Worker;
-import net.sf.jabref.logic.autosave.AutosaveManager;
-import net.sf.jabref.logic.autosave.BackupManager;
+import net.sf.jabref.logic.autosaveandbackup.AutosaveManager;
+import net.sf.jabref.logic.autosaveandbackup.BackupManager;
 import net.sf.jabref.logic.exporter.BibtexDatabaseWriter;
 import net.sf.jabref.logic.exporter.FileSaveSession;
 import net.sf.jabref.logic.exporter.SaveException;
