+++ src/main/java/net/sf/jabref/gui/importer/actions/OpenDatabaseAction.java
@@ -28,7 +28,7 @@ import net.sf.jabref.gui.autosave.BackupUIManager;
 import net.sf.jabref.gui.importer.ParserResultWarningDialog;
 import net.sf.jabref.gui.keyboard.KeyBinding;
 import net.sf.jabref.gui.shared.SharedDatabaseUIManager;
-import net.sf.jabref.logic.autosave.BackupManager;
+import net.sf.jabref.logic.autosaveandbackup.BackupManager;
 import net.sf.jabref.logic.importer.OpenDatabase;
 import net.sf.jabref.logic.importer.ParserResult;
 import net.sf.jabref.logic.l10n.Localization;
