+++ src/main/java/net/sf/jabref/model/database/event/AutosaveEvent.java
@@ -1,7 +1,7 @@
 package net.sf.jabref.model.database.event;
 
 /**
- * This Event is fired from {@link net.sf.jabref.logic.autosave.AutosaveManager} in case that a save task is pending.
+ * This Event is fired from {@link net.sf.jabref.logic.autosaveandbackup.AutosaveManager} in case that a save task is pending.
  */
 public class AutosaveEvent {
     // no data
